implemented tgt voltage change with buttons in modulation, also changed tgt type to be float so it is consistent everywhere, u_ref is also a global variable declared in system_params.h - testing required

+++ project_work/src/ui_control.c
@@ -30,7 +30,7 @@
 volatile SystemMode_t current_system_mode = MODE_CONFIG;
 
 // Degub variable:
-int tgt = 0;
+float tgt = 0;
 
 void ui_control_task(void *pvParameters){
